Refactor: Move OAuthInfoRepository and UserRepository to domain.user.dao
