+++ src/main/java/com/palindrome/studit/domain/user/application/UserService.java
@@ -4,8 +4,8 @@ import com.palindrome.studit.domain.user.entity.OAuthInfo;
 import com.palindrome.studit.domain.user.entity.OAuthProviderType;
 import com.palindrome.studit.domain.user.entity.User;
 import com.palindrome.studit.domain.user.entity.UserRoleType;
-import com.palindrome.studit.repository.OAuthInfoRepository;
-import com.palindrome.studit.repository.UserRepository;
+import com.palindrome.studit.domain.user.dao.OAuthInfoRepository;
+import com.palindrome.studit.domain.user.dao.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
